ga_controller/requirements.txt
pygad
pyyaml
requests
